Refactor PlayQueueInfo.Entry into its own class for reuse
